add additional dirty states to options
